requirements.txt
anthropic>=0.25.0
endee>=0.1.10
streamlit>=1.32.0
sentence-transformers>=2.6.0
python-dotenv>=1.0.0
